id type changed

+++ src/main/java/pl/kamil/rentcars/entity/Car.java
@@ -9,7 +9,7 @@ public class Car {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private int id;
+    private Long id;
 
     @Column(name = "brand")
     private String brand;
@@ -42,7 +42,7 @@ public class Car {
         this.fixing = fixing;
     }
 
-    public Car(int id, String brand, String model, String color, long vin, boolean rented, boolean fixing) {
+    public Car(Long id, String brand, String model, String color, long vin, boolean rented, boolean fixing) {
         this.id = id;
         this.brand = brand;
         this.model = model;
@@ -52,11 +52,11 @@ public class Car {
         this.fixing = fixing;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
